requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile
#
--extra-index-url https://download.pytorch.org/whl/cu121

aiohttp==3.9.1
    # via fsspec
aiosignal==1.3.1
    # via aiohttp
async-timeout==4.0.3
    # via aiohttp
attrs==23.1.0
    # via aiohttp
build==1.0.3
    # via pip-tools
certifi==2023.11.17
    # via requests
charset-normalizer==3.3.2
    # via requests
click==8.1.7
    # via pip-tools
filelock==3.13.1
    # via
    #   torch
    #   triton
frozenlist==1.4.1
    # via
    #   aiohttp
    #   aiosignal
fsspec[http]==2023.12.2
    # via
    #   lightning
    #   pytorch-lightning
    #   torch
idna==3.6
    # via
    #   requests
    #   yarl
importlib-metadata==7.0.0
    # via build
jinja2==3.1.2
    # via torch
lightning==2.1.3
    # via -r requirements.in
lightning-utilities==0.10.0
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
markupsafe==2.1.3
    # via jinja2
mpmath==1.3.0
    # via sympy
multidict==6.0.4
    # via
    #   aiohttp
    #   yarl
networkx==3.2.1
    # via torch
numpy==1.26.2
    # via
    #   lightning
    #   onnx
    #   onnxscript
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
onnx==1.15.0
    # via
    #   -r requirements.in
    #   onnxscript
onnxscript==0.1.0.dev20231230
    # via -r requirements.in
packaging==23.2
    # via
    #   build
    #   lightning
    #   lightning-utilities
    #   pytorch-lightning
    #   torchmetrics
pillow==10.1.0
    # via torchvision
pip-tools==7.3.0
    # via -r requirements.in
protobuf==4.25.1
    # via onnx
pyproject-hooks==1.0.0
    # via build
pytorch-lightning==2.1.3
    # via lightning
pyyaml==6.0.1
    # via
    #   lightning
    #   pytorch-lightning
requests==2.31.0
    # via
    #   fsspec
    #   torchvision
sympy==1.12
    # via torch
tomli==2.0.1
    # via
    #   build
    #   pip-tools
    #   pyproject-hooks
torch==2.1.1+cu121
    # via
    #   -r requirements.in
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==1.2.1
    # via
    #   lightning
    #   pytorch-lightning
torchvision==0.16.1+cu121
    # via -r requirements.in
tqdm==4.66.1
    # via
    #   lightning
    #   pytorch-lightning
triton==2.1.0
    # via torch
typing-extensions==4.9.0
    # via
    #   lightning
    #   lightning-utilities
    #   onnxscript
    #   pytorch-lightning
    #   torch
urllib3==2.1.0
    # via requests
wheel==0.42.0
    # via pip-tools
yarl==1.9.4
    # via aiohttp
zipp==3.17.0
    # via importlib-metadata

# The following packages are considered to be unsafe in a requirements file:
# pip
# setuptools
